Internals

Internals

DynaWAVE.WaveModel — Type.

WaveModel(G1::SparseMatrixCSC,G2::SparseMatrixCSC,
          meas::NetalignMeasure, seeds = [])

Internal function that creates a problem structure given two networks (static or dynamic) to align by and NetalignMeasure to optimize over. See align! on how perform the alignment.

Arguments

G1,G2 : input networks in sparse matrix format

meas : NetalignMeasure object (see NetalignMeasures.jl). WaveModel can optimize the DWEC measure (DWECMeasure), a dynamic network alignment measure and the WEC measure (WECMeasure), a static network alignment method.

source

DynaWAVE.align! — Method.

align!(M::WaveModel; maxiter::Integer=(size(M.G1,1)-length(M.seeds)),
            verbose=false) -> f

Internal function that performs alignment with respect to the parameters in WaveModel.

Arguments

maxiter : Stop aligning after maxiter iterations. Each iteration creates one aligned node pair.

verbose : Prints debugging outputs

source